authentication added to journal endpoints

+++ src/main/java/net/amit/journalApp/service/JournalEntryService.java
@@ -28,7 +28,7 @@ public class JournalEntryService {
             JournalEntry saved = journalEntryRepo.save(entry);
             user.getJournalEntries().add(saved);
 //            user.setUserName(null);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             throw new RuntimeException("An error occurred while saving the entry.",e);
         }
@@ -47,10 +47,19 @@ public class JournalEntryService {
         return journalEntryRepo.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepo.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed){
+                userService.saveUser(user);
+                journalEntryRepo.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while deleting the entry.",e);
+        }
+        return removed;
     }
 }
